requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
crispy-bootstrap5==0.7
cryptography==39.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.52.0
django-crispy-forms==2.0
django_debug_toolbar==4.0.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
psycopg==3.1.8
psycopg-binary==3.1.8
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.30.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==2.0.2
whitenoise==6.4.0
djangorestframework==3.14.0


